+++ src/OpenTelemetry.Exporter.Prometheus.AspNetCore/PrometheusExporterMiddleware.cs
@@ -4,6 +4,7 @@
 using System.Collections.Immutable;
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
+using System.Globalization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Net.Http.Headers;
 using OpenTelemetry.Exporter.Prometheus;
@@ -63,12 +64,23 @@ internal sealed class PrometheusExporterMiddleware
                 return;
             }
 
+            using var requestCancelled = new CancellationTokenSource();
+
+            if (TryGetScrapeTimeout(httpContext.Request.Headers, out var scrapeTimeout))
+            {
+                requestCancelled.CancelAfter(scrapeTimeout.GetValueOrDefault());
+            }
+
+            using var linkedCts = CancellationTokenSource.CreateLinkedTokenSource(requestCancelled.Token, httpContext.RequestAborted);
+
             var protocol = Negotiate(httpContext.Request);
 
             var collectionResponse = await this.exporter.CollectionManager.EnterCollect(protocol.IsOpenMetrics).ConfigureAwait(false);
 
             try
             {
+                linkedCts.Token.ThrowIfCancellationRequested();
+
                 var dataView = protocol.IsOpenMetrics ? collectionResponse.OpenMetricsView : collectionResponse.PlainTextView;
 
                 response.StatusCode = StatusCodes.Status200OK;
@@ -78,12 +90,24 @@ internal sealed class PrometheusExporterMiddleware
                     response.Headers.Append("Last-Modified", collectionResponse.GeneratedAtUtc.ToString("R"));
                     response.ContentType = PrometheusProtocol.GetContentType(protocol);
 
-                    await response.Body.WriteAsync(dataView.Array.AsMemory(0, dataView.Count)).ConfigureAwait(false);
+                    await response.Body.WriteAsync(dataView.Array.AsMemory(0, dataView.Count), linkedCts.Token).ConfigureAwait(false);
                 }
                 else
                 {
                     // It's not expected to have no metrics to collect, but it's not necessarily a failure, either.
                     PrometheusExporterEventSource.Log.NoMetrics();
+                }
+            }
+            catch (OperationCanceledException ex) when (ex.CancellationToken == linkedCts.Token)
+            {
+                if (scrapeTimeout is { } timeout)
+                {
+                    PrometheusExporterEventSource.Log.ScrapeTimedOut(timeout.TotalSeconds);
+                }
+
+                if (!response.HasStarted)
+                {
+                    response.StatusCode = StatusCodes.Status408RequestTimeout;
                 }
             }
             finally
@@ -98,6 +122,25 @@ internal sealed class PrometheusExporterMiddleware
             {
                 response.StatusCode = StatusCodes.Status500InternalServerError;
             }
+        }
+
+        static bool TryGetScrapeTimeout(
+            IHeaderDictionary headers,
+            [NotNullWhen(true)] out TimeSpan? scrapeTimeout)
+        {
+            const double MinTimeout = 0.001; // 1 millisecond
+            const double MaxTimeout = int.MaxValue / 1_000; // Prevent overflow of TimeSpan.FromSeconds()
+
+            if (headers.TryGetValue("X-Prometheus-Scrape-Timeout-Seconds", out var value) &&
+                double.TryParse(value, NumberStyles.AllowDecimalPoint, CultureInfo.InvariantCulture, out var scrapeTimeoutSeconds) &&
+                scrapeTimeoutSeconds is >= MinTimeout and <= MaxTimeout)
+            {
+                scrapeTimeout = TimeSpan.FromSeconds(scrapeTimeoutSeconds);
+                return true;
+            }
+
+            scrapeTimeout = null;
+            return false;
         }
     }